refactor: deduplicate fact display code into shared FactsPanel component
Extract duplicated FACT_CATEGORIES, getLatestFactsByProperty, groupByCategory,
and FactValueDisplay into shared FactsSection.tsx. Both org and person detail
pages now use the shared FactsPanel component, removing ~280 lines of duplication.
