+++ apps/web/src/app/organizations/[slug]/page.tsx
@@ -4,21 +4,20 @@ import { resolveOrgBySlug, getOrgSlugs } from "@/app/organizations/org-utils";
 import {
   getKBLatest,
   getKBFacts,
-  getKBEntity,
   getKBProperty,
-  getKBEntitySlug,
 } from "@/data/kb";
 import { getTypedEntityById, isOrganization } from "@/data";
 import {
-  formatKBFactValue,
   formatKBDate,
   titleCase,
   shortDomain,
-  isUrl,
 } from "@/components/wiki/kb/format";
-import type { Fact, Property } from "@longterm-wiki/kb";
 import Link from "next/link";
-import { Breadcrumbs } from "@/components/directory";
+import {
+  Breadcrumbs,
+  FactValueDisplay,
+  FactsPanel,
+} from "@/components/directory";
 
 export function generateStaticParams() {
   return getOrgSlugs().map((slug) => ({ slug }));
@@ -40,56 +39,6 @@ export async function generateMetadata({
 }
 
 // ── Subcomponents ─────────────────────────────────────────────────────
-
-/** Render a fact value, resolving ref/refs to entity name links. */
-function FactValueDisplay({ fact, property }: { fact: Fact; property?: Property }) {
-  const v = fact.value;
-  if (v.type === "ref") {
-    const refEntity = getKBEntity(v.value);
-    if (refEntity) {
-      const refSlug = getKBEntitySlug(v.value);
-      const href = refSlug && refEntity.type === "organization" ? `/organizations/${refSlug}`
-        : refSlug && refEntity.type === "person" ? `/people/${refSlug}`
-        : `/kb/entity/${v.value}`;
-      return (
-        <Link href={href} className="text-primary hover:underline">
-          {refEntity.name}
-        </Link>
-      );
-    }
-    return <span>{v.value}</span>;
-  }
-  if (v.type === "refs") {
-    return (
-      <span>
-        {v.value.map((refId, i) => {
-          const refEntity = getKBEntity(refId);
-          if (refEntity) {
-            const refSlug = getKBEntitySlug(refId);
-            const href = refSlug && refEntity.type === "organization" ? `/organizations/${refSlug}`
-              : refSlug && refEntity.type === "person" ? `/people/${refSlug}`
-              : `/kb/entity/${refId}`;
-            return (
-              <span key={refId}>
-                {i > 0 && ", "}
-                <Link href={href} className="text-primary hover:underline">
-                  {refEntity.name}
-                </Link>
-              </span>
-            );
-          }
-          return (
-            <span key={refId}>
-              {i > 0 && ", "}
-              {refId}
-            </span>
-          );
-        })}
-      </span>
-    );
-  }
-  return <span>{formatKBFactValue(fact, property?.unit, property?.display)}</span>;
-}
 
 function StatCard({
   label,
@@ -113,70 +62,6 @@ function StatCard({
       )}
     </div>
   );
-}
-
-/** Section header with optional count badge and divider. */
-function SectionHeader({ title, count }: { title: string; count?: number }) {
-  return (
-    <div className="flex items-center gap-3 mb-4">
-      <h2 className="text-base font-bold tracking-tight">{title}</h2>
-      {count != null && (
-        <span className="text-[11px] font-medium tabular-nums px-2 py-0.5 rounded-full bg-muted text-muted-foreground">
-          {count}
-        </span>
-      )}
-      <div className="flex-1 h-px bg-gradient-to-r from-border/60 to-transparent" />
-    </div>
-  );
-}
-
-// ── Fact sidebar helpers ──────────────────────────────────────────────
-
-const FACT_CATEGORIES: { id: string; label: string; order: number }[] = [
-  { id: "financial", label: "Financial", order: 0 },
-  { id: "product", label: "Products & Usage", order: 1 },
-  { id: "organization", label: "Organization", order: 2 },
-  { id: "safety", label: "Safety & Research", order: 3 },
-  { id: "people", label: "People", order: 4 },
-  { id: "other", label: "Other", order: 99 },
-];
-
-/** Group facts by property, taking only the latest per property. */
-function getLatestFactsByProperty(
-  facts: Fact[],
-): Map<string, Fact> {
-  const latest = new Map<string, Fact>();
-  for (const fact of facts) {
-    if (fact.propertyId === "description") continue;
-    if (!latest.has(fact.propertyId)) {
-      latest.set(fact.propertyId, fact);
-    }
-  }
-  return latest;
-}
-
-/** Group property IDs by category, returning sorted categories. */
-function groupByCategory(
-  propertyIds: string[],
-): Array<{ category: string; label: string; props: string[] }> {
-  const groups = new Map<string, string[]>();
-  for (const propId of propertyIds) {
-    const prop = getKBProperty(propId);
-    const category = prop?.category ?? "other";
-    const list = groups.get(category) ?? [];
-    list.push(propId);
-    groups.set(category, list);
-  }
-
-  const catMap = new Map(FACT_CATEGORIES.map((c) => [c.id, c]));
-  return [...groups.entries()]
-    .map(([catId, props]) => ({
-      category: catId,
-      label: catMap.get(catId)?.label ?? titleCase(catId),
-      order: catMap.get(catId)?.order ?? 99,
-      props,
-    }))
-    .sort((a, b) => a.order - b.order);
 }
 
 // ── Hero stat properties for org pages ────────────────────────────────
@@ -229,7 +114,7 @@ export default async function OrgProfilePage({
   // Header facts
   const hqFact = getKBLatest(entity.id, "headquarters");
 
-  // All facts for the sidebar
+  // All facts for the panel
   const allFacts = getKBFacts(entity.id).filter(
     (f) => f.propertyId !== "description",
   );
@@ -239,10 +124,6 @@ export default async function OrgProfilePage({
     : entity.wikiPageId
       ? `/wiki/${entity.wikiPageId}`
       : null;
-
-  // Fact sidebar data
-  const latestByProp = getLatestFactsByProperty(allFacts);
-  const categoryGroups = groupByCategory([...latestByProp.keys()]);
 
   // Description and website come from typed entity YAML data
   const descriptionText = orgData?.description ?? null;
@@ -345,46 +226,8 @@ export default async function OrgProfilePage({
       <div className="grid grid-cols-1 lg:grid-cols-3 gap-8">
         {/* Main content */}
         <div className="lg:col-span-2 space-y-8">
-          {/* Facts sidebar data displayed as main content when no records */}
           {allFacts.length > 0 && (
-            <section>
-              <SectionHeader title="Facts" count={latestByProp.size} />
-              <div className="border border-border/60 rounded-xl bg-card divide-y divide-border/40">
-                {categoryGroups.map(({ category, label, props }) => (
-                  <div key={category} className="px-4 py-3">
-                    <div className="text-[10px] font-semibold uppercase tracking-wider text-muted-foreground/60 mb-2">
-                      {label}
-                    </div>
-                    <div className="space-y-1.5">
-                      {props.map((propId) => {
-                        const fact = latestByProp.get(propId);
-                        if (!fact) return null;
-                        const property = getKBProperty(propId);
-                        return (
-                          <div
-                            key={propId}
-                            className="flex items-baseline justify-between gap-2 text-sm"
-                          >
-                            <span className="text-muted-foreground text-xs truncate">
-                              {property?.name ?? titleCase(propId)}
-                            </span>
-                            <span className="font-medium text-xs tabular-nums text-right shrink-0 max-w-[55%] truncate">
-                              <FactValueDisplay fact={fact} property={property} />
-                            </span>
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                ))}
-              </div>
-              <Link
-                href={`/kb/entity/${entity.id}`}
-                className="block mt-2 text-xs text-primary hover:underline text-center"
-              >
-                View all facts in KB explorer &rarr;
-              </Link>
-            </section>
+            <FactsPanel facts={allFacts} entityId={entity.id} />
           )}
         </div>
