+++ apps/web/src/components/directory/index.ts
@@ -4,3 +4,10 @@ export { CurrentBadge, FounderBadge } from "./StatusBadge";
 export { Breadcrumbs } from "./Breadcrumbs";
 export { SourceLink } from "./SourceLink";
 export { DirectoryEntityLink } from "./EntityLink";
+export {
+  FactValueDisplay,
+  FactsPanel,
+  getLatestFactsByProperty,
+  groupByCategory,
+  FACT_CATEGORIES,
+} from "./FactsSection";
